requirement.txt
torch
opencv-python
pytesseract
Pillow